+++ lib/src/types/message.dart
@@ -3,6 +3,7 @@
  * Copyright (c) 2020 Nebulino
  */
 
+/// Telegram Type Class: Message library
 library message;
 
 import 'package:json_annotation/json_annotation.dart';